requirements.txt
asgiref==3.2.3
Django==3.0.3
mysqlclient==1.4.6
Pillow==7.0.0
pytz==2019.3
sqlparse==0.3.0